dmamlogic new drivers and kernel
dts and dtsi files are for study only and won't apply so I will remove them later.

I took the defconfig from linux-meson64-v4.9.113-643134-ga8e83ea and changed it to be compatible with OV standards (what it has changed and what was not needed removed), the rest will be applied via linux-dmamlogic.inc

bluetooth module added to dm9x0 as it seems they needed it so we autoload it also.